refactor(image-vo): Remove URL validation
The url attribute will not always be an absolute url. In general, it will be interesting to add that base url in another place or at least, make it optional for the developer, to facilitate the use of CDNs and other issues. So we leave it as if it were a path.